Introduce CodeAwareTrait into Resource bundle
