+++ src/Sylius/Component/Addressing/Model/Zone.php
@@ -13,6 +13,7 @@ namespace Sylius\Component\Addressing\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Sylius\Component\Resource\Model\CodeAwareTrait;
 
 /**
  * @author Saša Stamenković <umpirsky@gmail.com>
@@ -20,15 +21,12 @@ use Doctrine\Common\Collections\Collection;
  */
 class Zone implements ZoneInterface
 {
+    use CodeAwareTrait;
+
     /**
      * @var mixed
      */
     protected $id;
-
-    /**
-     * @var string
-     */
-    protected $code;
 
     /**
      * @var string
@@ -69,22 +67,6 @@ class Zone implements ZoneInterface
     public function getId()
     {
         return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCode()
-    {
-        return $this->code;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCode($code)
-    {
-        $this->code = $code;
     }
 
     /**
