+++ src/Sylius/Component/Payment/Model/PaymentMethod.php
@@ -11,6 +11,7 @@
 
 namespace Sylius\Component\Payment\Model;
 
+use Sylius\Component\Resource\Model\CodeAwareTrait;
 use Sylius\Component\Resource\Model\TimestampableTrait;
 use Sylius\Component\Resource\Model\ToggleableTrait;
 use Sylius\Component\Translation\Model\AbstractTranslatable;
@@ -20,17 +21,12 @@ use Sylius\Component\Translation\Model\AbstractTranslatable;
  */
 class PaymentMethod extends AbstractTranslatable implements PaymentMethodInterface
 {
-    use TimestampableTrait, ToggleableTrait;
+    use CodeAwareTrait, TimestampableTrait, ToggleableTrait;
 
     /**
      * @var mixed
      */
     protected $id;
-
-    /**
-     * @var string
-     */
-    protected $code;
 
     /**
      * @var string
@@ -73,22 +69,6 @@ class PaymentMethod extends AbstractTranslatable implements PaymentMethodInterfa
     public function getId()
     {
         return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCode()
-    {
-        return $this->code;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCode($code)
-    {
-        $this->code = $code;
     }
 
     /**
