+++ src/Sylius/Component/Promotion/Model/Coupon.php
@@ -11,6 +11,7 @@
 
 namespace Sylius\Component\Promotion\Model;
 
+use Sylius\Component\Resource\Model\CodeAwareTrait;
 use Sylius\Component\Resource\Model\SoftDeletableTrait;
 use Sylius\Component\Resource\Model\TimestampableTrait;
 
@@ -19,17 +20,12 @@ use Sylius\Component\Resource\Model\TimestampableTrait;
  */
 class Coupon implements CouponInterface
 {
-    use SoftDeletableTrait, TimestampableTrait;
+    use CodeAwareTrait, SoftDeletableTrait, TimestampableTrait;
 
     /**
      * @var mixed
      */
     protected $id;
-
-    /**
-     * @var string
-     */
-    protected $code;
 
     /**
      * @var int
@@ -57,22 +53,6 @@ class Coupon implements CouponInterface
     public function getId()
     {
         return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCode()
-    {
-        return $this->code;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCode($code)
-    {
-        $this->code = $code;
     }
 
     /**
