+++ src/Sylius/Component/Promotion/Model/Promotion.php
@@ -13,6 +13,7 @@ namespace Sylius\Component\Promotion\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Sylius\Component\Resource\Model\CodeAwareTrait;
 use Sylius\Component\Resource\Model\SoftDeletableTrait;
 use Sylius\Component\Resource\Model\TimestampableTrait;
 
@@ -21,17 +22,12 @@ use Sylius\Component\Resource\Model\TimestampableTrait;
  */
 class Promotion implements PromotionInterface
 {
-    use SoftDeletableTrait, TimestampableTrait;
+    use CodeAwareTrait, SoftDeletableTrait, TimestampableTrait;
 
     /**
      * @var mixed
      */
     protected $id;
-
-    /**
-     * @var string
-     */
-    protected $code;
 
     /**
      * @var string
@@ -111,22 +107,6 @@ class Promotion implements PromotionInterface
     public function getId()
     {
         return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCode()
-    {
-        return $this->code;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCode($code)
-    {
-        $this->code = $code;
     }
 
     /**
