+++ src/Sylius/Component/Taxation/Model/TaxCategory.php
@@ -13,6 +13,7 @@ namespace Sylius\Component\Taxation\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Sylius\Component\Resource\Model\CodeAwareTrait;
 use Sylius\Component\Resource\Model\TimestampableTrait;
 
 /**
@@ -20,17 +21,13 @@ use Sylius\Component\Resource\Model\TimestampableTrait;
  */
 class TaxCategory implements TaxCategoryInterface
 {
-    use TimestampableTrait;
+    use CodeAwareTrait, TimestampableTrait;
 
     /**
      * @var mixed
      */
     protected $id;
 
-    /**
-     * @var string
-     */
-    protected $code;
     /**
      * @var string
      */
@@ -66,22 +63,6 @@ class TaxCategory implements TaxCategoryInterface
     public function getId()
     {
         return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCode()
-    {
-         return $this->code;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCode($code)
-    {
-        $this->code = $code;
     }
 
     /**
